Refactor some code model types
Removed the Generic type in Favor of new, more specific types Response
and Result.
Move handling of StdType and External after switch so that all types
being added are evaluated (no duplication).
Removed some discriminated types in favor of runtime checks in
constructors. It simplifies the calling code and centralizes the checks.