Emit TryFrom<T> trait for enum and model types
The trait is required only for types that are params and/or respones.
This requires contextual info on how the types are used. Added Context
type used by the code generator to make the determination.
Refactored emitter to use new CodeGenerator class which contains the
crate and its context.
AsyncMethod should have an array of MethodParameter types.
Fixed some bugs in the code model where kind wasn't being set.