Template for method bodies
This demonstrates what method bodies might look like. They're functional
and with them, the affected tests pass.
Changed the type of endpoint from String to azure_core::Url.
Added pipeline field to clients.
Added implementations for client constructor and hierarchical client
accessor methods. The sub-clients contain an immutable reference to the
endpoint and pipeline of the parent. This required adding a lifetime
annotation to the sub-clients.
Fixed response types for methods that return Unit.