Refactor some dependencies (#119)
* Refactor some dependencies

The tsp compiler and tcgc are now peer dependencies so the versions can
be controlled by the host. These are the only packages with symbols
imported by the emitter. Also included as dev dependencies.
The other dependencies are now dev dependencies as they're only need to
build the checked in tsp files.

* use ranges to support -dev testing