Add support for pageable methods
This only works with pageables that use the "nextLink" pattern.
Made some refinements to indentation so it's easier to consume and use
in unit tests.
buildIrBlock and buildMatch use the current indentation level and don't
add a trailing new-line at the end. This allows them to be nested within
other constructs.
Added optional returns field to matchArm.