Refactor creating mod.rs files
The creation of mod.rs files was inconsistent. For generated/mod.rs the
contents were pretty much hard-coded, and clients.mod.rs was smuggled in
with the client files which was kind of a hack.
This centralizes creation of mod.rs and makes their creation consitent.
It will also enable easier programmatic ways to add module declarations
which will be necessary in pending work for XML.