Refactor constructing operation paths
Use Url::join instead of Url::set_path as it will ensure that any
existing path on an endpoint is preserved. This required refactoring
constructing the Url in pageable methods to happen outside of the
closure (should be a net win as we need to clone less vars). It also
requires stripping off a leading forward slash on operations paths.
When an operation contains a single path parameter, directly join it
instead of creating a local and replacing it first.
For operations with a path of "/" omit setting it as it's redundant.
Make the name of the created Url var configurable.