Small refactoring of getting a param's value
Don't give fromSelf a default value as it makes it easy to inadvertently
miss passing the correct value. This was the case for setting optional
query parameters (currently no test coverage for this).
Clean up handling param name with potential "self." prefix. This makes
the handling uniform with less branches.