Improved support for supplemental paths
Client parameters in the code model are now a discriminated union of
endpoint and method params (similar to how header/path/query params are
handled in method params). This allows consumption of the params in the
correct places and removes the ambiguity.
Added cadl-ranch coverage for the newly supported functionality.
This is an addendum to 982c832.