Fix binary requests/responses
Requests that take a binary payload take a RequestContent<Bytes> param.
Responses that return a binary payload return a Response. Its generic
type parameter defaults to ResponseBody so it's omitted.
The code model was refactored so that RequestContent and Response model
types take a generic type parameter that better models what's
sent/received. This includes the Payload<T> type for operations that
send/receive a modeled type. It consolidates previous behavior of the
RequestContent and Response types (if we can send it, we can receive
it).