Refactor modules layout (#318)
* Refactor modules layout

Everything that's exported from generated/clients is now in that
directory (only method options was removed).
Everything that's exported from generated/models is now in a models
directory with the appropriate mod.rs file.
This has simplified the structure and the re-exporting of public surface
area (mod.rs files are much smaller now).
The merging of lib.rs content has been removed. Now, if a lib.rs file is
present it will not be overwritten and a warning emitted. Added the
switch overwrite-lib-rs to force overwriting (might be useful when
migrating to the new behavior).
Some model files were renamed to simplify exporting the models content.

* remove superfluous super:: when re-exporting clients

* use relative use path to models

* use crate::generated::models absolute path

simplify use statement for referencing sub-clients

* use crate::generated::clients instead of super