Consolidate use statements (#332)
* Consolidate use statements

The Use class now uses a tree to maintain the fully qualified path to an
import. This allows removal of some redundant import statements that
should have been nested to begin with.
The plural add methods on Use have been removed and the existing methods
are now variadic.

* refine handling of external types

renamed StdType to QualifiedType. External now extends QualifiedType as
they're essentially identical with the exception that External adds a
create dependency.

* omit serde for marker-only models

* don't export External type