Emit NoFormat for raw/empty responses
Split BodyFormat into PayloadFormat for requests and ResponseFormat for
responses as the set of values are slightly different.
ResponseFormat is now a required parameter when creating a Response.
Refactored some types to include the Response in their model so the
Format is propagated as required.
Removed some vestigial code from the Context class.