Fix some corner cases that get flagged by clippy
For pageable methods, only clone the API version parameter when there's
a paging strategy, else the cloned var isn't used.
For PollerState::More, don't make the request var mutable if the only
header is an optional Content-Type as it will be omitted.
Allow non snake-cased names for internal serde helpers for serializing
literal values.
Updated to the latest tcgc.