Naming and other fixes (#643)
Sanitize enum type names so invalid characters are removed. E.g. a tsp
union with the name foo-bar will create an enum named FooBar.
Don't allow the pascalCase helper to remove duplicate items in a
sequence as it can silently cause names to collide.
Allow clippy::too_many_arguments as required on method signatures.
Updated the tspcompile script to automatically use client.tsp when
available instead of explicitly specifying it (this is the cause for
some changes in the Spector test suite).